fix: Move AuditLogSchemaResponseModel to dto
