--- src/ai/backend/common/data/audit_log/BUILD
@@ -1,1 +1,0 @@
-python_sources(name="src")
